MTP-1738 Moved auto-accept logic to Check creation to accomodate displaying rules
